updating next and react
