+++ components/SamplesSection.js
@@ -1,5 +1,4 @@
 import React, { Component } from "react";
-import Slider from "react-slick";
 import Container from './Container';
 
 class Samples extends Component {
